Admin portal session and authorization.

+++ src/App/Controller/CategoryController.php
@@ -15,39 +15,64 @@ class CategoryController extends BaseController
     public function getAllCategories(RequestInterface $request, ResponseInterface $response, $args)
     {
         $this->logger->debug("Area:Category Action:getAllCategories Client:" . $_SERVER['REMOTE_ADDR']);
+        $siteDetail = $this->getSiteDetail();
+        $adminSession = $this->getAdminSession();
+        if ($adminSession == null || !$adminSession->getIsAdmin()) {
+            return $this->view->render($response, 'admin\accessdenied.html.twig', array('siteDetail' => $siteDetail, 'navSession' => $this->getNavSession('Access denied!', null, null), 'adminSession' => $adminSession));
+        }
         $categoriesMapper = new CategoryMapper($this->db);
         $categories = $categoriesMapper->readAll();      
-        return $this->view->render($response, 'admin\category.list.html.twig', array('siteDetail' => $this->getSiteDetail(), 'navSession' => $this->getNavSession('Categories', 'Categories', null), 'categories' => $categories));
+        return $this->view->render($response, 'admin\category.list.html.twig', array('siteDetail' => $siteDetail, 'navSession' => $this->getNavSession('Categories', 'Categories', null), 'adminSession' => $adminSession, 'categories' => $categories));
     }
 
     public function getCategoryForCreate(RequestInterface $request, ResponseInterface $response, $args)
     {
         $this->logger->debug("Area:Category Action:getCategoryForCreate Client:" . $_SERVER['REMOTE_ADDR']);
+        $siteDetail = $this->getSiteDetail();
+        $adminSession = $this->getAdminSession();
+        if ($adminSession == null || !$adminSession->getIsAdmin()) {
+            return $this->view->render($response, 'admin\accessdenied.html.twig', array('siteDetail' => $siteDetail, 'navSession' => $this->getNavSession('Access denied!', null, null), 'adminSession' => $adminSession));
+        }
         $data = array('id' => 0);
         $category = new Category($data);
-        return $this->view->render($response, 'admin\category.html.twig', array('siteDetail' => $this->getSiteDetail(), 'navSession' => $this->getNavSession('Add Category', 'Categories', null), 'category' => $category));
+        return $this->view->render($response, 'admin\category.html.twig', array('siteDetail' => $siteDetail, 'navSession' => $this->getNavSession('Add Category', 'Categories', null), 'adminSession' => $adminSession, 'category' => $category));
     }
 
     public function getCategoryById(RequestInterface $request, ResponseInterface $response, $args)
     {
         $this->logger->debug("Area:Category Action:getCategoryById Client:" . $_SERVER['REMOTE_ADDR'] . " Arguments:" . $args['id']);
+        $siteDetail = $this->getSiteDetail();
+        $adminSession = $this->getAdminSession();
+        if ($adminSession == null || !$adminSession->getIsAdmin()) {
+            return $this->view->render($response, 'admin\accessdenied.html.twig', array('siteDetail' => $siteDetail, 'navSession' => $this->getNavSession('Access denied!', null, null), 'adminSession' => $adminSession));
+        }
         $categoryMapper = new CategoryMapper($this->db);
         $category = $categoryMapper->read($args['id']);
-        return $this->view->render($response, 'admin\category.html.twig', array('siteDetail' => $this->getSiteDetail(), 'navSession' => $this->getNavSession('Edit Category', 'Categories', null), 'category' => $category));
+        return $this->view->render($response, 'admin\category.html.twig', array('siteDetail' => $siteDetail, 'navSession' => $this->getNavSession('Edit Category', 'Categories', null), 'adminSession' => $adminSession, 'category' => $category));
     }
 
     public function getCategoryByIdForDelete(RequestInterface $request, ResponseInterface $response, $args)
     {
         $this->logger->debug("Area:Category Action:getCategoryByIdForDelete Client:" . $_SERVER['REMOTE_ADDR'] . " Arguments:" . $args['id']);
+        $siteDetail = $this->getSiteDetail();
+        $adminSession = $this->getAdminSession();
+        if ($adminSession == null || !$adminSession->getIsAdmin()) {
+            return $this->view->render($response, 'admin\accessdenied.html.twig', array('siteDetail' => $siteDetail, 'navSession' => $this->getNavSession('Access denied!', null, null), 'adminSession' => $adminSession));
+        }
         $categoryMapper = new CategoryMapper($this->db);
         $categoryMapper->delete($args['id']);     
         $categories = $categoryMapper->readAll();        
-        return $this->view->render($response, 'admin\category.list.html.twig', array('siteDetail' => $this->getSiteDetail(), 'navSession' => $this->getNavSession('Categories', 'Categories', null), 'categories' => $categories));
+        return $this->view->render($response, 'admin\category.list.html.twig', array('siteDetail' => $siteDetail, 'navSession' => $this->getNavSession('Categories', 'Categories', null), 'adminSession' => $adminSession, 'categories' => $categories));
     }
 
     public function postCategory(RequestInterface $request, ResponseInterface $response, $args)
     {
         $this->logger->debug("Area:Category Action:postCategory Client:" . $_SERVER['REMOTE_ADDR']);
+        $siteDetail = $this->getSiteDetail();
+        $adminSession = $this->getAdminSession();
+        if ($adminSession == null || !$adminSession->getIsAdmin()) {
+            return $this->view->render($response, 'admin\accessdenied.html.twig', array('siteDetail' => $siteDetail, 'navSession' => $this->getNavSession('Access denied!', null, null), 'adminSession' => $adminSession));
+        }
         $data = $request->getParsedBody();
         $data['slug'] = str_replace(' ', '-', strtolower($data['name']));
         if (array_key_exists ('published', $data)) {
@@ -69,6 +94,6 @@ class CategoryController extends BaseController
             $id = $categoryMapper->create($category);
             $category->setId($id);
         }
-        return $this->view->render($response, 'admin\category.html.twig', array('siteDetail' => $this->getSiteDetail(), 'navSession' => $this->getNavSession('Edit Categories', 'Categories', null), 'category' => $category));     
+        return $this->view->render($response, 'admin\category.html.twig', array('siteDetail' => $siteDetail, 'navSession' => $this->getNavSession('Edit Categories', 'Categories', null), 'adminSession' => $adminSession, 'category' => $category));     
     }
 }
